Refactored code and added class styles
