+++ src/Components/PokemonList/pokemon.css
@@ -1,0 +1,24 @@
+
+thead tr {
+    background-color: #a8cef3;
+}
+
+.tableContainer {
+    max-width: 400px;
+    margin: auto;
+    border-radius: 0;
+    box-shadow: none;
+    border: 1px solid #ccc;
+}
+.pagination-footer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #e0ecfb;
+    padding: 1;
+}
+                
+
+
+
+
